Baseconvert receives an unsigned long int

+++ holberton.h
@@ -66,7 +66,7 @@ char *conv_non_printable(char *s);
 void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size);
 void itohex_2bytes(unsigned int n, char *dest);
 char *_itobi(unsigned int n, char *buff, int size);
-char *base_convert(char*, int, unsigned int, int, int);
+char *base_convert(char*, int, unsigned long int, int, int);
 char *getAddress_p(void*, char*, int);
 int putInBuffer(char *buff, int *b_cnt, char *s, int len);
 void checkFlags(char *buff, int *b_cnt, flag *flags, int *new_buffs);
